Move RFC to Edward
