+++ src/Edward/proxy_supervisor.cpp
